refactor cocktails into single page
