+++ src/components/cocktailbox.tsx
@@ -1,7 +1,19 @@
-import React from "react";
 import Image from "next/image";
-import { renderRatingMoons } from "@/components/renderratingmoons";
 import styles from "@/styles/cocktails.module.css";
+
+// Function to convert numeric rating to moons
+// (was stars but the half star emoji doesn't exist so I am being creative)
+const renderRatingMoons = (rating: number) => {
+  const fullMoons = Math.floor(rating);
+  const remaining = rating - fullMoons;
+  const hasHalfMoon = remaining >= 0.25 && remaining < 0.75;
+  const emptyMoons = 5 - fullMoons - (hasHalfMoon ? 1 : 0);
+  const moonString =
+    `🌝`.repeat(fullMoons) +
+    (hasHalfMoon ? `🌗` : "") +
+    `🌚`.repeat(emptyMoons);
+  return <span>{moonString}</span>;
+};
 
 // Create the design for each cocktail entry on the cocktails page
 interface Cocktail {
